+++ NEWS.md
@@ -3,6 +3,8 @@
 - Add delay in loop to wait for service principal during AKS resource creation; could timeout prematurely otherwise.
 - `KubernetesCluster$create()`, `apply()`, etc now accept HTTP\[S\] URLs as well as filenames as arguments.
 - Use the processx package to run external commands, rather than `base::system2()`. A major benefit of this change is that command output is automatically captured and returned as an R object, making it easier to write automated scripts.
+  - The commandline is now a list component of the R object, rather than an attribute.
+- The various `DockerRegistry` and `KubernetesCluster` methods for calling docker, kubectl and helm now have `...` as an argument, allowing you to pass extra inputs to these commands as needed.
 
 # AzureContainers 1.1.2